build(android): update target and compile SDK to 34
Hardcode the Android compileSdk and targetSdk versions to 34, and minSdk to 21. This provides more explicit control over the Android build environment.

This update also resolves a namespace issue with dependencies, allowing for the removal of the temporary `sed` script from the CI release workflow that was patching the `image_gallery_saver` package.